Add Storage, TaskList, Ui classes
